+++ src/main/java/frc/robot/Robot.java
@@ -40,10 +40,10 @@ import frc.robot.Constants.ControllerK;
 import frc.robot.Constants.DriveK;
 import frc.robot.Constants.IntakeK;
 import frc.robot.Constants.SwerveK;
+import frc.robot.Field.Cage;
+import frc.robot.Field.ReefSide;
 import frc.robot.commands.Autos;
 import frc.robot.commands.Routines;
-import frc.robot.commands.Routines.Cage;
-import frc.robot.commands.Routines.ReefSide;
 import frc.robot.lib.logging.LogUtil;
 import frc.robot.lib.logging.TalonFXLogger;
 import frc.robot.lib.util.DriveUtil;
